[REVIEW][kernels] address batch-invariant logprob comments
